# api/docs/swagger.yaml
definitions:
  template_variables.CreateResponse:
    properties:
      id:
        type: string
    type: object
...
  template_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  user_service.GetAllOrganizationResponse:
    properties:
      count:
        type: integer
      project:
        items:
          $ref: '#/definitions/user_service.Organization'
        type: array
      simple:
        items:
          $ref: '#/definitions/user_service.Organization'
        type: array
    type: object
  user_service.GetAllOrganizationsResponse:
    properties:
      count:
        type: integer
      organizations:
        items:
          $ref: '#/definitions/user_service.Organization'
        type: array
    type: object
...
  user_service.Organization:
    properties:
      code:
        type: integer
      created_at:
        type: string
      description:
        type: string
      external_id:
        type: integer
      full_name:
        type: string
      id:
        type: string
      inn:
        type: string
      name:
        type: string
      responsible_orgs:
        items:
          $ref: '#/definitions/user_service.ResponsibleUser'
        type: array
      soato:
        type: integer
      status:
        type: boolean
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  user_service.ResponsibleUser:
    properties:
      organization_id:
        type: string
      soato:
        type: string
    type: object
...
  var_user_service.CreateUpdateOrganizationSwag:
    properties:
      code:
        type: string
      description:
        type: string
      external_id:
        type: integer
      full_name:
        type: string
      id:
        type: string
      inn:
        type: string
      name:
        type: string
      responsible_orgs:
        items:
          $ref: '#/definitions/var_user_service.ResponsibleUser'
        type: array
      soato:
        type: integer
      status:
        type: boolean
      type:
        type: string
    type: object
...
  var_user_service.ResponsibleUser:
    properties:
      organization_id:
        type: string
      soato:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /v1/organization:
    get:
      consumes:
      - application/json
      description: API for getting all organizations
      parameters:
      - description: type
        in: query
        name: type
        type: string
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetAllOrganizationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Getting All Organizations
      tags:
      - organization
    post:
      consumes:
      - application/json
      description: API for creating organization
      parameters:
      - description: organization
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/var_user_service.CreateUpdateOrganizationSwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/template_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Organization
      tags:
      - organization
  /v1/organization/{organization_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: organization_id
        in: path
        name: organization_id
        required: true
        type: string
      - description: inn
        in: query
        name: inn
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Organization'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Get Organization
      tags:
      - organization
    put:
      consumes:
      - application/json
      description: API for updating organization
      parameters:
      - description: organization_id
        in: path
        name: organization_id
        required: true
        type: string
      - description: organization
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/var_user_service.CreateUpdateOrganizationSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Update Organization
      tags:
      - organization
  /v1/organizations/dashboard:
    get:
      consumes:
      - application/json
      description: API for getting all organizations
      parameters:
      - description: type
        in: query
        name: type
        type: string
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetAllOrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Getting All Organizations
      tags:
      - organization
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"